[nanoleaf] Remove the device type configuration setting.
Autodetect it instead.

The device type setting was used to turn on support for touch gestures.
Since we know the controller model, we already know if touch is supported.

The setting was also confusing since it let you select light panels or
canvas, but the binding also works for shapes, elements and lines.

Signed-off-by: Jørgen Austvik <jaustvik@acm.org>